seuil minimum , nb entier

+++ src/main/java/afg/achat/afgApprovAchat/model/Article.java
@@ -22,7 +22,7 @@ public class Article {
     String codeArticle;
     String designation;
     @Column(name = "seuil_min")
-    double seuilMin;
+    int seuilMin;
     @ManyToOne @JoinColumn(name = "id_udm" , referencedColumnName = "id_udm")
     Udm udm;
     @ManyToOne @JoinColumn(name = "id_famille" , referencedColumnName = "id_famille")
@@ -30,7 +30,7 @@ public class Article {
     @ManyToOne @JoinColumn(name = "id_centre_budgetaire" , referencedColumnName = "id_centre_budgetaire")
     CentreBudgetaire centreBudgetaire;
 
-    public Article(String designation, double seuilMin, Udm udm, Famille famille, CentreBudgetaire centreBudgetaire) {
+    public Article(String designation, int seuilMin, Udm udm, Famille famille, CentreBudgetaire centreBudgetaire) {
         this.setDesignation(designation);
         this.setSeuilMin(seuilMin);
         this.setUdm(udm);
